remove export of old files
