+++ src/components/IndexGroupsList.tsx
@@ -1,81 +1,97 @@
 import React from "react"
 import { useEffect, useState } from "react"
-import styled from "styled-components"
+
 import axios from "axios"
 
-//hooks
-import useApplicationData from "./hooks/useApplicationData"
-import usePublic from "./hooks/usePublic"
+function GroupTestElement(props) {
 
-//components
-import IndexGroupsListItem from "./IndexGroupsListItem"
+   const { name, id } = props.group
+   const userId = JSON.parse(localStorage.getItem("session") || "{}")
+   .id
+   const [sub, setSub] = useState(false)
+   const [disable, setDisable] = useState(false)
+   console.log('sub true or not', props.sub)
+   console.log("from grouptestelent", name)
+   console.log("from goruptestele", id)
+   console.log('-------------')
 
-//helpers
-import subscribeList from "../helpers/selectors"
+   useEffect(()=>{
+      axios({
+         method: 'get',
+         url: `http://localhost:3001/group/${id}/${userId}`
+      }).then(isSubbed => {
+         setSub(isSubbed.data)
+      })
+   },[])
+   
+   const handleSub = () => {
+     
+      const data = {userId}
+      sub ? (
+         axios({
+            method: "delete",
+            url: `http://localhost:3001/group/subscription/delete/${id}`,
+            data: data,
+         })
+         .then(() => {
+            console.log('unsub success')
+         })
+         .catch(() => console.log('unsub unsuccess'))
+      )
+      : (
+         axios({
+            method: "post",
+            url: `http://localhost:3001/group/subscription/${id}`,
+            data: data,
+         })
+         .then(() => {
+            console.log('sub success')
+            
+         })
+         .catch(() => console.log('sub unsuccess') )   
+      )
+      setSub(!sub)
+      setDisable(true)
+   }
+   
+   return (
+      <div>
+         <p key={id}>{name}</p>
+         <button onClick={handleSub}>{sub ? '-' : '+'}</button>
+      </div>
+   )
 
+}
 
-export default function IndexGroupsList() {
-
-  //state includes list of users groups state= { groups{[id:number ,name:string], ...} }
-  const { state } = useApplicationData()
-  const { publicGroups } = usePublic()
-  const unsubscribe = "-"
-  const toSubscribe = "+"
-
-  // const [groupState, setGroupState] = useState({})
-  const subscribeListResult = subscribeList(state, publicGroups)
-
-  const groupsArr = publicGroups.groups
+export default function IndexGroupList() {
+  const [allGroups, setAllGroups] = useState([])
+//   const [groupBelong, setGroupBelong] = useState([])
   useEffect(() => {
-    // groupsArr.forEach(group => {
-    //   if (subscribeListResult.includes(group.id)) {
-    //     setGroupState(prev => ({ ...prev, [group.id]: true }))
-    //   } else {
-    //     setGroupState(prev => ({ ...prev, [group.id]: false }))
-    //   }
-    // })
-  }, []) //[groupsArr.length, subscribeListResult.length])
-
-  const onSubmitFunction = (event: any, groupId: number, button: string) => {
-    event.preventDefault()
     const userId: number = JSON.parse(localStorage.getItem("session") || "{}")
       .id
-    const data = { userId }
+    console.log("userID", userId)
+    axios.get("http://localhost:3001/group/public").then(data => {
+      console.log(data.data)
+      setAllGroups(data.data)
+    })
+//     axios.get(`http://localhost:3001/group/u/${userId}`).then(data => {
+//       const groupNameBelongArr = data.data.map(group => {
+//          return group.name
+//        })
+//        console.log("groupArr", groupNameBelongArr)
+//       setGroupBelong(groupNameBelongArr)
+//     })
+  }, [])
 
-    if (button === unsubscribe) {
-      axios({
-        method: "delete",
-        url: `http://localhost:3001/group/subscription/delete/${groupId}`,
-        data: data,
-      }).then(res => {
-        // console.log(res.data)
-      })
-    }
-    if (button === toSubscribe) {
-      axios({
-        method: "post",
-        url: `http://localhost:3001/group/subscription/${groupId}`,
-        data: data,
-      }).then(res => {
-        // console.log(res.data)
-      })
-    }
-  }
-
-  const groupsList = groupsArr.map((group, index :number) => {
-    let button = toSubscribe
-    if (subscribeListResult.includes(group.id)) {
-      button = unsubscribe
-    }
-
-    return (
-      <IndexGroupsListItem
-        key={index}
-        group={group}
-        button={button}
-        onSubmitFunction={onSubmitFunction}
-      />
-    )
+  const List = allGroups.map(group => {
+      // if (groupBelong.includes(group.name)) {
+      //    return <GroupTestElement group={group} sub={true} />
+      // } else {
+      //   
+      // }
+      console.log('group =>' , group)
+      return <GroupTestElement group={group} />
   })
-  return <div>{groupsList}</div>
+
+  return <div>{List}</div>
 }
