+++ src/components/TopicBoard.tsx
@@ -5,8 +5,7 @@ import MuiExpansionPanel from "@material-ui/core/ExpansionPanel"
 import MuiExpansionPanelSummary from "@material-ui/core/ExpansionPanelSummary"
 import MuiExpansionPanelDetails from "@material-ui/core/ExpansionPanelDetails"
 import Typography from "@material-ui/core/Typography"
-import IndexGroupsList from "./IndexGroupsList"
-import Test from './test'
+import IndexGroupList from './IndexGroupsList'
 const ExpansionPanel = withStyles({
   root: {
     border: "1px solid rgba(0, 0, 0, .125)",
@@ -111,7 +110,7 @@ export default function TopicBoard({  }: Props) {
           <Typography>Groups</Typography>
         </ExpansionPanelSummary>
 
-          <Test />
+          <IndexGroupList />
 
         <ExpansionPanelDetails>
           <Typography></Typography>
