+++ src/helpers/.selectors.tsx
@@ -16,4 +16,4 @@ const subscribeList = (state, publicGroups) => {
 
   return subscribeListArr
 }
-export default subscribeList
+// export default subscribeList
